+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.13.40 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- WEB-3415 : Authorize (add) Link type in Plone Site root (and upgrade step).
+  [boulch]
 
 
 0.13.39 (2020-08-24)
@@ -2495,4 +2496,3 @@ Changelog
 ----------------
 
 - Initial release
-